packages/prompts/pyproject.toml
[project]
name = "prompts"
version = "0.1.0"
description = "Versioned prompt templates for anime-rag"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "langchain>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prompts"]
